@@ -1,6 +1,6 @@
 source "https://rubygems.org"
 
-gem "rails", "3.2.13"
+gem "rails", "4.0.0"
 
 gem "email_signup", path: "engines/email_signup"
 gem "teaser", path: "engines/teaser"
