@@ -9,20 +9,20 @@
 # from scratch. The latter is a flawed and unsustainable approach (the more migrations
 # you'll amass, the slower it'll run and the greater likelihood for issues).
 #
-# It's strongly recommended to check this file into your version control system.
+# It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(:version => 20130403220851) do
+ActiveRecord::Schema.define(version: 20130403220851) do
 
-  create_table "email_signup_entries", :force => true do |t|
+  create_table "email_signup_entries", force: true do |t|
     t.string   "email"
-    t.datetime "created_at", :null => false
-    t.datetime "updated_at", :null => false
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
   end
 
-  create_table "event_counter_counts", :force => true do |t|
+  create_table "event_counter_counts", force: true do |t|
     t.string  "object_identifier"
     t.string  "event_identifier"
-    t.integer "count",             :default => 0
+    t.integer "count",             default: 0
   end
 
 end
