@@ -4,7 +4,7 @@ require File.expand_path('../boot', __FILE__)
 # require "active_record/railtie"
 require "action_controller/railtie"
 require "action_mailer/railtie"
-require "active_resource/railtie"
+# require "active_resource/railtie"
 require "sprockets/railtie"
 #require "rails/test_unit/railtie"
 
@@ -67,5 +67,7 @@ module TheNextBigThing
     config.assets.version = '1.0'
 
     #config.assets.initialize_on_precompile = false
+    
+    config.secret_key_base = "some super secret"
   end
 end
