@@ -3,7 +3,11 @@ module EmailSignup
     isolate_namespace EmailSignup
 
     initializer :append_migrations do |app|
-      app.config.paths["db/migrate"] += config.paths["db/migrate"].expanded unless app.root.to_s.match root.to_s
+      unless app.root.to_s.match root.to_s
+        config.paths["db/migrate"].expanded.each do |path|
+          app.config.paths["db/migrate"] << path 
+        end
+      end
     end
   end
 end
